Move common host building to separate file.
