+++ src/Tests/Microsoft.Diagnostics.Monitoring.OpenApiGen/Program.cs
@@ -41,9 +41,8 @@ namespace Microsoft.Diagnostics.Monitoring.OpenApiGen
             // Create all of the same services as dotnet-monitor and add
             // OpenAPI generation in order to have it inspect the ASP.NET Core
             // registrations and descriptions.
-            IHost host = DiagnosticsMonitorCommandHandler
+            IHost host = HostBuilderHelper
                 .CreateHostBuilder(
-                    console: null,
                     urls: Array.Empty<string>(),
                     metricUrls: Array.Empty<string>(),
                     metrics: true,
